test: address review gaps in regression coverage
Move compile-only checks into colocated type-test files so typechecking and portable-runtime checks own them. Strengthen context dependency validation, restore omitted-name rejection coverage, and verify notifier failures preserve successful exit status.

Update the audit to reflect these fixes and the parent PR's documentation test removals. Production behavior is unchanged.
