+++ packages/core/src/parsing/schema.test.ts
@@ -2,8 +2,6 @@ import { describe, expect, it } from "bun:test";
 
 import type { StandardSchema } from "@crustjs/utils/schema";
 
-import type { Equal, Expect } from "../../tests/helpers.ts";
-
 type StandardInput = Parameters<StandardSchema["~standard"]["validate"]>[0];
 
 import { defineExtension } from "../api/extension.ts";
@@ -227,15 +225,3 @@ describe("schema interaction with Extensions", () => {
 		expect(validated).toBe(false);
 	});
 });
-
-// Compile-time regression checks; intentionally never invoked.
-// the schema output type reaches the Command Action
-function _typecheckTheSchemaOutputTypeReachesTheCommandAction() {
-	new Crust("cli")
-		.args({ name: "port", schema: port() })
-		.flags({ name: "tag", type: "string", schema: port() })
-		.action((_ctx) => {
-			type _argOutput = Expect<Equal<(typeof _ctx.args)["port"], number>>;
-			type _flagOutput = Expect<Equal<(typeof _ctx.flags)["tag"], number>>;
-		});
-}
