+++ packages/core/src/validation/direct-names.brands.test.ts
@@ -27,17 +27,3 @@ describe("direct definition name brands", () => {
 		expect(renameCommand(dynamicName).name).toBe("dynamic");
 	});
 });
-
-// Compile-time regression checks; intentionally never invoked.
-// rejects statically known empty names at every direct definition API
-function _typecheckRejectsStaticallyKnownEmptyNamesAtEveryDirectDefinitionAPI() {
-	const command = defineCommand("valid", (builder) => builder);
-	// @ts-expect-error -- flag names must be non-empty
-	defineFlag("", { type: "string" });
-	// @ts-expect-error -- argument names must be non-empty
-	defineArg("", { type: "string" });
-	// @ts-expect-error -- command names must be non-empty
-	defineCommand("", (builder) => builder);
-	// @ts-expect-error -- renamed command names must be non-empty
-	command.as("");
-}
